+++ app/test/e2e/endpoint.spec.js
@@ -64,7 +64,6 @@ describe('GET Endpoints', () => {
                 binary: false,
                 cache: [],
                 uncache: [],
-                toDelete: false,
                 path: '/v1/dataset',
                 method: 'POST',
                 pathRegex: {},