+++ app/test/e2e/test.constants.js
@@ -104,7 +104,6 @@ const endpointTest = {
     binary: false,
     cache: [],
     uncache: [],
-    toDelete: false,
     path: '/v1/dataset',
     method: 'POST',
     pathRegex: '',