+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # 14/02/2020
+- Remove `toDelete` from microservice model.
 - Add `createdAt` and `updatedAt` fields to the endpoint model.
 - Add basic filters to get endpoints endpoint.